+++ src/main/java/com/grupo3/Caso1/Dao/Posgrest/DetalleGarantiaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.grupo3.Caso1.Model.DetalleGarantia;
 
-public interface DetalleGarantiaRepositorio extends JpaRepository<DetalleGarantia, Integer> {
+public interface DetalleGarantiaRepositorio extends JpaRepository<DetalleGarantia, Long> {
 
 }
