+++ src/main/java/com/grupo3/Caso1/Dao/Posgrest/GaratiaVehiculoRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.grupo3.Caso1.Model.GarantiaVehiculo;
 import com.grupo3.Caso1.Model.Vehiculo;
 
-public interface GaratiaVehiculoRepositorio extends JpaRepository<GarantiaVehiculo, Integer> {
+public interface GaratiaVehiculoRepositorio extends JpaRepository<GarantiaVehiculo, Long> {
 
 }
